Create Issue #62 Account Creation Bug
This is not a code fix, but is a pull request to acknowledge the finding of several bugs during the account creation screen. https://github.com/ayastreb/money-tracker/issues/62 Contains the actual description of the bug. This is simply to follow the pull request procedure for the submittal of a bug or supposed bug.